feat: enable button new and removes old button
