+++ .storybook/storybook.requires.js
@@ -33,7 +33,6 @@ const getStories = () => {
     require("../src/elements/Banner/Banner.stories.tsx"),
     require("../src/elements/Box/Box.stories.tsx"),
     require("../src/elements/Button/Button.stories.tsx"),
-    require("../src/elements/ButtonNew/Button.stories.tsx"),
     require("../src/elements/Checkbox/Checkbox.stories.tsx"),
     require("../src/elements/Collapse/Collapse.stories.tsx"),
     require("../src/elements/CollapsibleMenuItem/CollapsibleMenuItem.stories.tsx"),
