+++ src/elements/Button/Button.tsx
@@ -1,35 +1,43 @@
-import { TextVariant } from "@artsy/palette-tokens/dist/typography/v3" // TODO: remove palette-tokens when this file (Button.tsx) is removed.
-import { useState } from "react"
-import { PressableProps, TextStyle, GestureResponderEvent, Pressable } from "react-native"
+import { useEffect, useState } from "react"
+import { PressableProps, GestureResponderEvent, Pressable } from "react-native"
 import Haptic, { HapticFeedbackTypes } from "react-native-haptic-feedback"
-import { config } from "react-spring"
-// @ts-ignore
-import { animated, Spring } from "react-spring/renderprops-native"
-import styled from "styled-components/native"
-import { Color, SpacingUnit } from "../../types"
-import { useColor } from "../../utils/hooks"
+import Animated, {
+  interpolateColor,
+  useAnimatedReaction,
+  useAnimatedStyle,
+  useDerivedValue,
+  useSharedValue,
+  withTiming,
+} from "react-native-reanimated"
+import { useColorsForVariantAndState } from "./colors"
+import { MeasuredView, ViewMeasurements } from "../../elements/MeasuredView"
 import { Box, BoxProps } from "../Box"
-import { Flex } from "../Flex/Flex"
-import { MeasuredView, ViewMeasurements } from "../MeasuredView"
+import { Flex } from "../Flex"
 import { Spacer } from "../Spacer"
 import { Spinner } from "../Spinner"
-import { Text } from "../Text/Text"
-import { useTextStyleForPalette } from "../Text/helpers"
+import { Text, useTextStyleForPalette } from "../Text"
+
+const ANIMATION_DURATION = 150
+
+type ButtonSize = "small" | "large"
+type ButtonVariant =
+  | "fillDark"
+  | "fillLight"
+  | "fillGray"
+  | "fillSuccess"
+  | "outline"
+  | "outlineGray"
+  | "outlineLight"
+  | "text"
 
 export interface ButtonProps extends BoxProps {
   children: React.ReactNode
 
-  size?: "small" | "large"
-  variant?:
-    | "fillDark"
-    | "fillLight"
-    | "fillGray"
-    | "fillSuccess"
-    | "outline"
-    | "outlineGray"
-    | "outlineLight"
-    | "text"
+  size?: ButtonSize
+  variant?: ButtonVariant
   onPress?: PressableProps["onPress"]
+  onPressIn?: PressableProps["onPressIn"]
+  onPressOut?: PressableProps["onPressOut"]
 
   icon?: React.ReactNode
   iconPosition?: "left" | "left-start" | "right"
@@ -56,80 +64,68 @@ export interface ButtonProps extends BoxProps {
   longestText?: string
 
   /** Used only for tests and stories */
-  testOnly_state?: DisplayState
-
-  textVariant?: TextVariant
+  testOnly_pressed?: PressableProps["testOnly_pressed"]
 }
 
-enum DisplayState {
-  Enabled = "enabled",
-  Disabled = "disabled",
-  Loading = "loading",
-  Pressed = "pressed",
-}
-
-export const Button: React.FC<ButtonProps> = ({
+export const Button = ({
   children,
-  disabled,
+  disabled: disabledProp,
   haptic,
   icon,
   iconPosition = "left",
-  loading,
+  loading: loadingProp,
+  block,
   longestText,
   onPress,
+  onPressIn,
+  onPressOut,
   size = "large",
   variant = "fillDark",
-  testOnly_state,
+  testOnly_pressed,
   testID,
-  textVariant,
   hitSlop,
-  ...rest
-}) => {
-  const textVariantBySize = size === "small" ? "xs" : "sm"
-  const textStyle = useTextStyleForPalette(textVariant ?? textVariantBySize)
+  ...restProps
+}: ButtonProps) => {
+  const [disabled] = useStateWithProp(!!disabledProp)
+  const [loading] = useStateWithProp(!!loadingProp)
+  const [_, setPressed, pressedV] = useStateWithProp(!!testOnly_pressed)
 
-  const [innerDisplayState, setInnerDisplayState] = useState(DisplayState.Enabled)
+  const pressAnimationProgress = useSharedValue(0)
+  useAnimatedReaction(
+    () => {
+      return pressedV.value
+    },
+    (pressedVal) => {
+      return (pressAnimationProgress.value = withTiming(pressedVal, {
+        duration: ANIMATION_DURATION,
+      }))
+    }
+  )
 
+  const textStyle = useTextStyleForPalette(size === "small" ? "xs" : "sm")
   const [longestTextMeasurements, setLongestTextMeasurements] = useState<ViewMeasurements>({
     width: 0,
     height: 0,
   })
 
-  const displayState =
-    testOnly_state ?? // if we use the test prop, use that
-    (loading // if we have loading or disabled in props, they are used
-      ? DisplayState.Loading
-      : disabled
-      ? DisplayState.Disabled
-      : innerDisplayState) // otherwise use the inner state for pressed or enabled
-
-  const getSize = (): { height: number; mx: SpacingUnit } => {
+  const height = (() => {
     switch (size) {
       case "small":
-        return { height: 30, mx: "15px" }
+        return 30
       case "large":
-        return { height: 50, mx: "30px" }
+        return 50
     }
-  }
+  })()
+
+  const spinnerColor = variant === "text" ? "blue100" : "white100"
 
   const handlePress = (event: GestureResponderEvent) => {
     if (onPress === undefined || onPress === null) {
       return
     }
 
-    if (displayState === DisplayState.Loading || displayState === DisplayState.Disabled) {
+    if (disabled || loading) {
       return
-    }
-
-    // Did someone tap really fast? Flick the highlighted state
-    if (displayState === DisplayState.Enabled) {
-      setInnerDisplayState(DisplayState.Pressed)
-      setTimeout(() => {
-        setInnerDisplayState(DisplayState.Enabled)
-      }, 0.3)
-    } else {
-      // Was already selected
-      setInnerDisplayState(DisplayState.Enabled)
     }
 
     if (haptic !== undefined) {
@@ -139,365 +135,154 @@ export const Button: React.FC<ButtonProps> = ({
     onPress(event)
   }
 
-  const containerSize = getSize()
-  const to = useStyleForVariantAndState(variant, testOnly_state ?? displayState)
+  const colorsForVariantAndState = useColorsForVariantAndState()
+
+  const containerColorsAnim = useAnimatedStyle(() => {
+    const colors = colorsForVariantAndState[variant]
+    if (disabled) {
+      return {
+        backgroundColor: colors.disabled.bg,
+        borderColor: colors.disabled.border,
+      }
+    }
+    return {
+      backgroundColor: interpolateColor(
+        pressAnimationProgress.value,
+        [0, 1],
+        [colors.active.bg, colors.pressed.bg]
+      ),
+      borderColor: interpolateColor(
+        pressAnimationProgress.value,
+        [0, 1],
+        [colors.active.border, colors.pressed.border]
+      ),
+    }
+  })
+
+  const textAnim = useAnimatedStyle(() => {
+    const colors = colorsForVariantAndState[variant]
+    if (loading) {
+      return { color: "rgba(0, 0, 0, 0)" }
+    }
+
+    return {
+      color: interpolateColor(
+        pressAnimationProgress.value,
+        [0, 1],
+        [colors.active.text, colors.pressed.text]
+      ),
+      textDecorationLine: pressAnimationProgress.value > 0 ? "underline" : "none",
+    }
+  })
 
   return (
-    <Spring native to={to} config={config.stiff}>
-      {(springProps: typeof to) => (
-        <Pressable
-          hitSlop={hitSlop}
-          testOnly_pressed={testOnly_state === DisplayState.Pressed}
-          disabled={testOnly_state === DisplayState.Disabled || disabled}
-          onPressIn={() => {
-            if (displayState === DisplayState.Loading) {
-              return
-            }
-            setInnerDisplayState(DisplayState.Pressed)
-          }}
-          onPressOut={() => {
-            if (displayState === DisplayState.Loading) {
-              return
-            }
-            setInnerDisplayState(DisplayState.Enabled)
-          }}
-          onPress={handlePress}
-          testID={testID}
+    <Pressable
+      disabled={disabled}
+      onPressIn={(e) => {
+        onPressIn?.(e)
+        setPressed(true)
+      }}
+      onPressOut={(e) => {
+        onPressOut?.(e)
+        setPressed(false)
+      }}
+      onPress={handlePress}
+      testID={testID}
+      testOnly_pressed={testOnly_pressed}
+      hitSlop={hitSlop}
+    >
+      <Flex flexDirection="row">
+        <Flex
+          {...restProps}
+          height={height}
+          width={block ? "100%" : undefined}
+          borderRadius={50}
+          overflow="hidden"
         >
-          <Flex flexDirection="row">
-            <AnimatedContainer
-              {...rest}
-              style={{
-                backgroundColor: springProps.backgroundColor,
-                borderColor: springProps.borderColor,
-                height: containerSize.height,
-              }}
+          <AFlex borderWidth={1} borderRadius={50} overflow="hidden" style={containerColorsAnim}>
+            <Flex
+              height="100%"
+              mx="25px"
+              flexDirection="row"
+              alignItems="center"
+              justifyContent="center"
             >
-              <Flex mx={containerSize.mx}>
-                <Flex height="100%" flexDirection="row" alignItems="center" justifyContent="center">
-                  {iconPosition === "left-start" && !!icon ? (
-                    <Box position="absolute" left={0}>
-                      {icon}
-                      <Spacer x={0.5} />
-                    </Box>
-                  ) : null}
-                  {iconPosition === "left" && !!icon ? (
-                    <>
-                      {icon}
-                      <Spacer x={0.5} />
-                    </>
-                  ) : null}
-                  {/* This makes sure that in testing environment the button text is
-                      not rendered twice, in normal environment this is not visible.
-                      This will result in us being able to use getByText over
-                      getAllByText()[0] to select the buttons in the test environment.
-                  */}
-                  {!__TEST__ && longestText && (
-                    <MeasuredView setMeasuredState={setLongestTextMeasurements}>
-                      <Text color="red" style={textStyle}>
-                        {longestText ? longestText : children}
-                      </Text>
-                    </MeasuredView>
-                  )}
-                  <AnimatedText
-                    style={[
-                      {
-                        width: longestText ? Math.ceil(longestTextMeasurements.width) : "auto",
-                        color: springProps.textColor,
-                        textDecorationLine: springProps.textDecorationLine,
-                      },
-                      textStyle,
-                    ]}
-                    textAlign="center"
-                  >
-                    {children}
-                  </AnimatedText>
-                  {iconPosition === "right" && !!icon && (
-                    <>
-                      <Spacer x={0.5} />
-                      {icon}
-                    </>
-                  )}
-                </Flex>
+              {iconPosition === "left-start" && !!icon ? (
+                <Box position="absolute" left={0}>
+                  {icon}
+                  <Spacer x={0.5} />
+                </Box>
+              ) : null}
 
-                {displayState === DisplayState.Loading && (
-                  <SpinnerContainer>
-                    <Spinner size={size} color={to.loaderColor} />
-                  </SpinnerContainer>
-                )}
-              </Flex>
-            </AnimatedContainer>
-          </Flex>
-        </Pressable>
-      )}
-    </Spring>
+              {iconPosition === "left" && !!icon ? (
+                <>
+                  {icon}
+                  <Spacer x={0.5} />
+                </>
+              ) : null}
+
+              <AText
+                style={[{ width: Math.ceil(longestTextMeasurements.width) }, textStyle, textAnim]}
+                textAlign="center"
+                selectable={false}
+              >
+                {children}
+              </AText>
+
+              <MeasuredView setMeasuredState={setLongestTextMeasurements}>
+                <Text color="red" style={textStyle}>
+                  {longestText ? longestText : children}
+                </Text>
+              </MeasuredView>
+
+              {iconPosition === "right" && !!icon ? (
+                <>
+                  <Spacer x={0.5} />
+                  {icon}
+                </>
+              ) : null}
+
+              {loading ? (
+                <Box
+                  position="absolute"
+                  width="100%"
+                  height="100%"
+                  alignItems="center"
+                  justifyContent="center"
+                >
+                  <Spinner size={size} color={spinnerColor} />
+                </Box>
+              ) : null}
+            </Flex>
+          </AFlex>
+        </Flex>
+      </Flex>
+    </Pressable>
   )
 }
 
-const useStyleForVariantAndState = (
-  variant: Exclude<ButtonProps["variant"], undefined>,
-  state: DisplayState
-): {
-  backgroundColor: string
-  borderColor: string
-  borderWidth?: number
-  textColor: string
-  loaderColor: Color
-  textDecorationLine?: TextStyle["textDecorationLine"]
-} => {
-  const color = useColor()
+const AText = Animated.createAnimatedComponent(Text)
+const AFlex = Animated.createAnimatedComponent(Flex)
 
-  const retval = {
-    textDecorationLine: "none",
-  } as ReturnType<typeof useStyleForVariantAndState>
+const useStateWithProp = (
+  prop: boolean
+): [
+  boolean,
+  React.Dispatch<React.SetStateAction<boolean>>,
+  Readonly<Animated.SharedValue<1 | 0>>
+] => {
+  const [state, setState] = useState(!!prop)
 
-  switch (variant) {
-    case "fillDark":
-      retval.textColor = color("white100")
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("black100")
-          retval.borderColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("black30")
-          retval.borderColor = color("black30")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("black100")
-          retval.borderColor = color("black100")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "white100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
+  useEffect(() => {
+    setState(!!prop)
+  }, [prop])
 
-    case "fillLight":
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("white100")
-          retval.textColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("black30")
-          retval.borderColor = color("black30")
-          retval.textColor = color("white100")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("white100")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "black100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = color("white100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
+  const stateV = useDerivedValue(() => {
+    if (!!state) {
+      return 1
+    }
+    return 0
+  }, [state])
 
-    case "fillGray":
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("black10")
-          retval.borderColor = color("black10")
-          retval.textColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("black30")
-          retval.borderColor = color("black30")
-          retval.textColor = color("white100")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("black10")
-          retval.borderColor = color("black10")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "black100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = color("white100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    case "fillSuccess":
-      retval.textColor = color("white100")
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "white100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue10")
-          retval.borderColor = color("blue10")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    case "outline":
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black60")
-          retval.textColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black30")
-          retval.textColor = color("black30")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black60")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "black100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = color("white100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    case "outlineGray":
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black30")
-          retval.textColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black30")
-          retval.textColor = color("black30")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = color("white100")
-          retval.borderColor = color("black30")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "black100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = color("white100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    case "outlineLight":
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.backgroundColor = "rgba(0, 0, 0, 0)"
-          retval.borderColor = color("white100")
-          retval.textColor = color("white100")
-          break
-        case DisplayState.Disabled:
-          retval.backgroundColor = "rgba(0, 0, 0, 0)"
-          retval.borderColor = color("black30")
-          retval.textColor = color("black30")
-          break
-        case DisplayState.Loading:
-          retval.backgroundColor = "rgba(0, 0, 0, 0)"
-          retval.borderColor = color("white100")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "white100"
-          break
-        case DisplayState.Pressed:
-          retval.backgroundColor = color("blue100")
-          retval.borderColor = color("blue100")
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    case "text":
-      retval.backgroundColor = "rgba(0, 0, 0, 0)"
-      retval.borderColor = "rgba(0, 0, 0, 0)"
-      switch (state) {
-        case DisplayState.Enabled:
-          retval.textColor = color("black100")
-          break
-        case DisplayState.Disabled:
-          retval.textColor = color("black30")
-          break
-        case DisplayState.Loading:
-          retval.textColor = "rgba(0, 0, 0, 0)"
-          retval.loaderColor = "blue100"
-          break
-        case DisplayState.Pressed:
-          retval.textColor = color("blue100")
-          retval.textDecorationLine = "underline"
-          break
-        default:
-          null
-      }
-      break
-
-    default:
-      null
-  }
-
-  return retval
+  return [state, setState, stateV]
 }
-
-const Container = styled(Box)<ButtonProps>`
-  position: relative;
-  border-width: 1px;
-  border-radius: 50px;
-  width: ${(p) => (p.block ? "100%" : "auto")};
-  overflow: hidden;
-`
-
-const SpinnerContainer = styled(Box)<ButtonProps>`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  align-items: center;
-  justify-content: center;
-`
-
-const AnimatedContainer = animated(Container)
-const AnimatedText = animated(Text)
-
-export { DisplayState as _test_DisplayState }
