--- src/elements/Button/index.ts
@@ -1,4 +1,0 @@
-export * from "./Button"
-export * from "./CTAButton"
-export * from "./LinkButton"
-export * from "./FollowButton"
