+++ src/elements/Button/index.tsx
@@ -1,4 +1,3 @@
 export * from "./Button"
 export * from "./CTAButton"
 export * from "./FollowButton"
-export * from "./LinkButton"
