--- src/elements/ButtonNew/CTAButton.tsx
@@ -1,3 +1,0 @@
-import { Button, ButtonProps } from "./Button"
-
-export const CTAButton = (props: ButtonProps) => <Button block haptic {...props} />
