--- src/elements/ButtonNew/LinkButton.tsx
@@ -1,8 +1,0 @@
-import { Text, TextProps } from "../Text"
-import { Touchable } from "../Touchable"
-
-export const LinkButton = (props: TextProps) => (
-  <Touchable onPress={props.onPress}>
-    <Text underline {...props} />
-  </Touchable>
-)
